•	A single page blogging application for fast user interaction
Edit, update, insert and delete of posts and comments in real-time. MongoDB for database is used which is a NoSQL DB. Node.js as backend, ember as front-end and MongoDB as NoSQL database made the application faster for the end-user. There is a folder name S2G8-blog-backend which has the code for the back-end